backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
spotipy==2.23.0
pandas==2.2.2
mysql-connector-python==9.0.0
python-dotenv==1.0.1
pydantic==2.9.0
